--- backend/src/features/definitions/income/infrastructure/entity.ts
@@ -1,17 +1,0 @@
-import { z } from 'zod'
-
-import { IncomeDefinitionKind } from '../domains/entity'
-
-export const IncomeDefinitionRecord = z.object({
-  id: z.string(),
-  user_id: z.string(),
-  name: z.string(),
-  kind: z.enum(IncomeDefinitionKind),
-  value: z.number(),
-  is_taxable: z.number(),
-  enabled_at: z.number(),
-  disabled_at: z.number(),
-  updated_at: z.number(),
-})
-
-export type IncomeDefinitionRecord = z.infer<typeof IncomeDefinitionRecord>
